+++ CHANGELOG.md
@@ -101,6 +101,7 @@
 - Removed the two deprecated functions from Session.php !1259 @koenvg
 - Extend RegionPermissions to a mayAdministrateRegions() method. Removes the topbar menu entry if false !1236 @jofranz
 - Update to PHP 7.4.5 and fix some backwards incompatible changes @NerdyProjects
+- Add database constraints to fs_faq !1436 @ffm_hessen
 
 ## Dev/Test/CI stuff
 - Add "linux" tag for finding CI servers !1332 @nicksellen